Create jest setup script file
This is a Jest setup script file, which is used to configure and set up the testing environment for a React application. Jest is a popular testing framework for JavaScript applications, and this file is used to define settings and configurations that are applied before running tests.

Here's a breakdown of the file:

1. module.exports: The file exports a function that returns an object with setup configurations.
2. setupFilesAfterEnv: This configuration specifies files that should be executed after the environment has been set up.
3. <rootDir>/node_modules/react-native/jest/setup.js: This file is a setup script provided by React Native to configure the testing environment.
4. <rootDir>/jest.setup.js: This file is another setup script that is specific to the project.

The file is essentially setting up the testing environment by executing the specified setup scripts. These scripts may perform tasks such as:

- Configuring mock modules
- Setting up mock APIs
- Defining global variables
- Configuring enzyme or other testing libraries

By setting up the testing environment, this file ensures that tests are executed consistently and reliably, and that the application is properly configured for testing.